refactor: Refactor routes
