+++ src/routers/login/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import LoginContainer from "../components/login/login-container";
+import LoginContainer from "../../components/login/login-container";
 
 const Login = () => {
   return <LoginContainer />;
