+++ src/routers/volunteer/create/styles.ts
@@ -1,0 +1,12 @@
+/* eslint-disable import/prefer-default-export */
+import styled from "@emotion/styled";
+
+export const BodyContainer = styled.div`
+  > * {
+    margin: 16px 0;
+  }
+
+  margin: 16px auto;
+  max-width: 1280px;
+  width: 100%;
+`;
