+++ src/routers/volunteer/index.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+import Footer from "../../components/common/footer";
+import Header from "../../components/common/header";
+import HeaderTopSpace from "../../components/common/header-top-space";
+import VolunteerList from "../../components/volunteer-list";
+
+const Volunteers = () => {
+  return (
+    <>
+      <Header />
+      <HeaderTopSpace />
+      <VolunteerList />
+      <Footer />
+    </>
+  );
+};
+
+export default Volunteers;
